HBASE-29614 Remove static final field modification in tests around Unsafe (#7337)
(cherry picked from commit e0cec314c839158eac7dbd080582ebce83718a0b)
